% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspr.R
\name{.CheckRooted}
\alias{.CheckRooted}
\title{Check that trees are rooted}
\usage{
.CheckRooted(trees)
}
\arguments{
\item{trees}{A \code{phylo} object or a list of \code{phylo} objects.}
}
\value{
Called for its side-effect (stops with an error if any tree is
unrooted).
}
\description{
Check that trees are rooted
}
\keyword{internal}
