% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspr.R
\name{.PrepareTrees}
\alias{.PrepareTrees}
\title{Prepare trees for passing to uspr}
\usage{
.PrepareTrees(
  tree1,
  tree2,
  allPairs = FALSE,
  checks = TRUE,
  keepLabels = FALSE
)
}
\arguments{
\item{keepLabels}{Logical specifying whether to pass text labels to distance
calculator.  This is slower, but makes maximum agreement forests easier
to read.}
}
\value{
A two-element list, each entry of which is a character vector listing
trees in Newick format.
}
\description{
Converts trees to Newick strings, as expected by the \code{uspr} library.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
