% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspr.R
\name{TreeRearrangementDistances}
\alias{TreeRearrangementDistances}
\alias{USPRDist}
\alias{ReplugDist}
\alias{TBRDist}
\title{Tree Rearrangement Distances}
\usage{
USPRDist(tree1, tree2, checks = TRUE)

ReplugDist(tree1, tree2, checks = TRUE, maf = FALSE)

TBRDist(tree1, tree2, checks = TRUE, maf = FALSE, printMafs = FALSE,
  countMafs = FALSE, optimize = TRUE, protectB = TRUE,
  exact = FALSE, approximate = !exact, approxEstimate = TRUE,
  tbrEstimate = TRUE)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, or lists thereof.}

\item{checks}{Logical specifying whether to check that trees are properly
formatted and labelled.  Specify \code{FALSE} at your peril, as improper
input is likely to crash R.}

\item{maf}{Logical specifying whether to report a maximum agreement forest
corresponding to the optimal score.}
}
\description{
Tree Rearrangement Distances
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
