% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspr-package.R
\docType{package}
\name{uspr-package}
\alias{uspr}
\alias{uspr-package}
\title{uspr: Calculate SPR and TBR distance between unrooted phylogenetic trees}
\description{
Core functions for the creation, modification and analysis of 
  phylogenetic trees.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/uspr}
  \item \url{https://github.com/ms609/uspr}
  \item \url{https://github.com/cwhidden/uspr}
  \item Report bugs at \url{https://github.com/ms609/uspr/issues}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (0000-0001-5660-1727) [copyright holder]

Other contributors:
\itemize{
  \item Chris Whidden \email{cwhidden@fredhutch.org} [copyright holder]
}

}
\keyword{internal}
