% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspr.R
\name{TreeRearrangementDistances}
\alias{TreeRearrangementDistances}
\alias{USPRDist}
\alias{ReplugDist}
\alias{TBRDist}
\title{Calculate Tree Rearrangement Distances}
\usage{
USPRDist(tree1, tree2, checks = TRUE, useTbrApproxEstimate = TRUE,
  useTbrEstimate = TRUE, useReplugEstimate = TRUE)

ReplugDist(tree1, tree2, checks = TRUE, maf = FALSE)

TBRDist(tree1, tree2, checks = TRUE, exact = FALSE,
  approximate = !exact, maf = FALSE, countMafs = FALSE,
  printMafs = FALSE, optimize = TRUE, protectB = TRUE)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, or lists thereof.}

\item{checks}{Logical specifying whether to check that trees are properly
formatted and labelled.  Specify \code{FALSE} at your peril, as improper
input is likely to crash R.}

\item{useTbrApproxEstimate, useTbrEstimate, useReplugEstimate}{Logical specifying
whether to use approximate TBR distance, TBR distance or Replug distance to
help estimate the SPR distance.}

\item{maf}{Logical specifying whether to report a maximum agreement forest
corresponding to the optimal score.}

\item{exact}{Logical specifying whether to calculate the exact TBR distance.}

\item{approximate}{Logical specifying whether to calculate the approximate
TBR distance.  By default, is set to the opposite of \code{exact}; either
\code{approximate} or \code{exact} should usually be set to \code{TRUE} if a distance is
required.}

\item{countMafs}{Logical specifying whether to count the number of Maximum
Agreement Forests found.}

\item{printMafs}{Logical specifying whether to print Maximum Agreement
Forests to stdout whilst counting.
Use \code{\link{capture.output}}\code{(TBRDist(tree1, tree2, printMafs = TRUE))} to access
these in R.}

\item{optimize}{Logical specifying whether to use the default optimizations.}

\item{protectB}{Logical specifying whether to use the PROTECT_B optimization.
Overrides value inherited from \code{optimize}.}
}
\value{
\code{USPRDist} returns a vector of SPR distances between each pair of
unrooted trees.

\code{ReplugDist} returns a vector of Replug distances between each pair
of trees, or (if \code{maf = TRUE}) a named list whose second and third elements
list a vector of maximum agreement forests for each pair of trees.

\code{TBRDist} returns a named list, each element of which bears a vector
corresponding to the requested value for each tree pair.  If only the exact
value is requested (\code{exact = TRUE}), an unnamed vector of distances is
returned.
}
\description{
Calculate Tree Rearrangement Distances
}
\examples{
  tree1 <- TreeTools::BalancedTree(9)
  tree2 <- TreeTools::PectinateTree(9)

  # SPR distance
  USPRDist(tree1, tree2)


  # Replug distance
  ReplugDist(tree1, tree2)
  ReplugDist(tree1, tree2, maf = TRUE)

  # TBR distance between two trees
  TBRDist(tree1, tree2, exact = TRUE)

  # Compare a list against one tree, using approximate distances
  TBRDist(list(tree1, tree2), tree2, exact = FALSE)

  # Compare two lists
  TBRDist(list(tree1, tree2, tree2),
          list(tree2, tree1, tree2),
          exact = TRUE, approximate = TRUE, countMafs = TRUE)

  # Capture maximum agreement forests
  mafs <- capture.output(TBRDist(tree1, tree2, approximate = FALSE,
                          printMafs = TRUE))
  head(mafs)

}
\references{
If you use \emph{uspr} in your research, please cite:
\itemize{
\item Chris Whidden and Frederick A. Matsen IV. Calculating the Unrooted
Subtree-Prune-and-Regraft Distance. eprint
\href{http://arxiv.org/abs/1511.07529}{arXiv:1511.07529}.
}
}
\author{
Algorithm by Chris Whidden (\href{mailto:cwhidden@fredhutch.org}{cwhidden@fredhutch.org})

R wrappers by Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
